refactor detr_post_process
