+++ easycv/models/detection/detectors/detr/detr_head.py
@@ -1,14 +1,11 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 # Copyright (c) Alibaba, Inc. and its affiliates.
-import numpy as np
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 
 from easycv.models.builder import HEADS, build_neck
-from easycv.models.detection.utils import (HungarianMatcher, SetCriterion,
-                                           box_cxcywh_to_xyxy,
-                                           box_xyxy_to_cxcywh)
+from easycv.models.detection.utils import (DetrPostProcess, HungarianMatcher,
+                                           SetCriterion, box_xyxy_to_cxcywh)
 from easycv.models.utils import MLP
 
 
@@ -49,7 +46,7 @@ class DETRHead(nn.Module):
             weight_dict=weight_dict,
             eos_coef=eos_coef,
             losses=['labels', 'boxes'])
-        self.postprocess = PostProcess()
+        self.postprocess = DetrPostProcess()
         self.transformer = build_neck(transformer)
 
         self.class_embed = nn.Linear(embed_dims, num_classes + 1)
@@ -148,42 +145,4 @@ class DETRHead(nn.Module):
         orig_target_sizes = torch.stack(ori_shape_list, dim=0)
 
         results = self.postprocess(outputs, orig_target_sizes, img_metas)
-        return results
-
-
-class PostProcess(nn.Module):
-    """ This module converts the model's output into the format expected by the coco api"""
-
-    @torch.no_grad()
-    def forward(self, outputs, target_sizes, img_metas):
-        """ Perform the computation
-        Parameters:
-            outputs: raw outputs of the model
-            target_sizes: tensor of dimension [batch_size x 2] containing the size of each images of the batch
-                          For evaluation, this must be the original image size (before any data augmentation)
-                          For visualization, this should be the image size after data augment, but before padding
-        """
-        out_logits, out_bbox = outputs['pred_logits'], outputs['pred_boxes']
-
-        assert len(out_logits) == len(target_sizes)
-        assert target_sizes.shape[1] == 2
-
-        prob = F.softmax(out_logits, -1)
-        scores, labels = prob[..., :-1].max(-1)
-
-        # convert to [x0, y0, x1, y1] format
-        boxes = box_cxcywh_to_xyxy(out_bbox)
-        # and from relative [0, 1] to absolute [0, height] coordinates
-        img_h, img_w = target_sizes.unbind(1)
-        scale_fct = torch.stack([img_w, img_h, img_w, img_h],
-                                dim=1).to(boxes.device)
-        boxes = boxes * scale_fct[:, None, :]
-
-        results = {
-            'detection_boxes': [boxes[0].cpu().numpy()],
-            'detection_scores': [scores[0].cpu().numpy()],
-            'detection_classes': [labels[0].cpu().numpy().astype(np.int32)],
-            'img_metas': img_metas
-        }
-
         return results
