+++ easycv/models/detection/utils/__init__.py
@@ -5,8 +5,8 @@ from .boxes import (batched_nms, bbox2result, bbox_overlaps, bboxes_iou,
                     generalized_box_iou, postprocess)
 from .generator import MlvlPointGenerator
 from .matcher import HungarianMatcher
-from .misc import (accuracy, filter_scores_and_topk, fp16_clamp,
-                   gen_encoder_output_proposals, gen_sineembed_for_position,
-                   interpolate, inverse_sigmoid, output_postprocess,
-                   select_single_mlvl)
+from .misc import (DetrPostProcess, accuracy, filter_scores_and_topk,
+                   fp16_clamp, gen_encoder_output_proposals,
+                   gen_sineembed_for_position, interpolate, inverse_sigmoid,
+                   output_postprocess, select_single_mlvl)
 from .set_criterion import SetCriterion
